Script for grabbing MITOCW links
This script gives all the available .zip links for each course from Electrical and Computer Science from the MIT OCW site. Later it can be all copied to clipboard and pasted in a downloader or open each link separately. Change the links accordingly to get to the required department and sections.